display background, walls, borders, collectibles, exit, player(to solve)

+++ librairies/libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: twang <twang@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 12:41:50 by twang             #+#    #+#             */
-/*   Updated: 2023/01/13 17:21:29 by twang            ###   ########.fr       */
+/*   Updated: 2023/02/20 12:33:17 by twang            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s && s[i])
